Netowrds: Proof of concept (one partial round)
Committing an Atari BASIC listing (as ASCII, translated to
ATASCII EOL characters via a simple `tr` invocation in Makefile)
that plays part of a round of the game.

Updated the server-side PHP scripts to fix some bugs and
improve things, and address Fujinet's current inability to handle
"Transfer-Encoding: chunked", and do Unix LF -> ATASCII EOL
conversion on the server side.